// src/app/toolbar/model/toolbarelement.model.ts
export class element {
  element_id?: any;
  element_name?: any;
  element_data?: toolelementData;
  element_type?: any;
  element_uuid?: any;
  is_hidden?: any;
  is_removed?: any;
  version_number?: any;
  element_category?: any;
  placeholderNumber?: any;
  _id?: any;
  name?: any;
  id: any;
}
export class toolelementData {
  element_id?: any;
  element_type?: any;
  element_name?: any;
  element_category?: any;
  element_order?: any;
  element_label?: any;
  stroke_color?: any;
  fill_color?: any;
  line_width?: any;
  opacity?: any;
  element_uuid: any;
}